moved nav to it's own folder
