+++ client/src/components/Nav/Nav.jsx
@@ -1,5 +1,5 @@
 import { Link, useNavigate } from "react-router-dom";
-import useAuth from "../hooks/useAuth";
+import useAuth from "../../hooks/useAuth";
 import "./Nav.css";
 
 const Nav = () => {
